# docs/swagger.yaml
basePath: /api
definitions:
  dto.CreateAllocationRequest:
    properties:
      allocation_type:
        enum:
        - PERCENTAGE
        - NOMINAL
        type: string
      nominal:
        type: number
      pocket_id:
        type: string
      priority:
        maximum: 3
        minimum: 1
        type: integer
      user_platform_id:
        type: string
    required:
    - allocation_type
    - nominal
    - priority
    type: object
...
  dto.UpdateAllocationRequest:
    properties:
      allocation_type:
        enum:
        - PERCENTAGE
        - NOMINAL
        type: string
      is_active:
        type: boolean
      nominal:
        type: number
      pocket_id:
        type: string
      priority:
        maximum: 3
        minimum: 1
        type: integer
      user_platform_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive financial management system with smart allocation engine,
    transaction tracking, and detailed reports
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Coin Backend API
  version: "1.0"
paths:
...
  /v1/allocations:
    get:
      description: Get all allocations for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Allocations retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all allocations
      tags:
      - Allocations
    post:
      consumes:
      - application/json
      description: Create a new salary allocation rule
      parameters:
      - description: Allocation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAllocationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Allocation created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new allocation
      tags:
      - Allocations
  /v1/allocations/{id}:
    delete:
      description: Delete an allocation
      parameters:
      - description: Allocation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Allocation deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete allocation
      tags:
      - Allocations
    get:
      description: Get a specific allocation by ID
      parameters:
      - description: Allocation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Allocation retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get allocation by ID
      tags:
      - Allocations
    put:
      consumes:
      - application/json
      description: Update an allocation
      parameters:
      - description: Allocation ID
        in: path
        name: id
        required: true
        type: string
      - description: Update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAllocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Allocation updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update allocation
      tags:
      - Allocations
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"